fix size formatting on model version artifacts table
